requirements.txt
asgiref==3.7.2
boto3==1.34.12
botocore==1.34.12
Django==5.0
django-storages==1.14.2
djangorestframework==3.14.0
jmespath==1.0.1
mysqlclient==2.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.7
